requirements.txt
streamlit
pandas
numpy
scikit-learn
joblib
Pillow
requests
xgboost
tensorflow